requirements/base.txt
Django>=5.0,<5.3
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
channels>=4.0
channels-redis>=4.2
psycopg2-binary>=2.9
django-cors-headers>=4.4
pyotp>=2.9
qrcode>=7.4
Pillow>=10.3
daphne>=4.1
django-jazzmin>=2.6.0
django-cors-headers==4.7.0
django-filter>=24.2
